Refactor: Extract ElasticsearchBase class
